Prototype coloring, reduce window resizing errors, refactor stuff
V-EZ has the audacity to auto-resize the swapchain image (which then may not match the framebuffer). I can only mitigate this by checking for a resize immediately before the queuePresent, but it can still happen sometimes. I am currently working on a replacement for V-EZ which will hopefully fix the issue (among others).